Initial test setup, ready to try a few fixes
